refactor(ajax): Move things around a little and add more documentation
